Simplify CSS rules and improve print styles
I've replaced use of `rem` with `em`, as it simplifies the style rules.

I also noticed some weird behaviour with print styles where, if the `html` element's `font-size` was less than 100%, it would reduce the width of the printed page. For example, with `html { font-size: 50%; }`, the printed page would only fill the left half of the page. To rectify this, I've removed all style rules on the `html` element. They're no longer necessary since I've adjusted the styles to use `em` instead of `rem` units.